Update char_count_DB.json
"user_id" would be where a new account's ID is added. It's key is the total characters entered by that person. "user_id_tmc" is the total message count of that account since they were 'registered' by the bot. Users will be added in automatically. As soon as a message from a new user is detected, they are added.