Typing paragraphs and specifications
"typeconfig.py" contains paragraphs and their specifications. So far, there are three descriptive paragraphs on the Hobbit. Two short, and one long.
"[cog1]commands.py" will list all the commands in the future. I meant to put it in the cogs folder, but it seemed to be raising errors there, so I just moved it here for now.