0.2.0-beta (#33)
* Working addprompt command

* Get Prompts (#27)

* get prompts command

* ingore classpath

* TP Rebalance (#28)

Changed the TP calculation to be more fair towards higher WPMs. Previously, the TP gain from WPM was linear, but it is now exponential.
The curve for the length bonus in TR has been changed for prompts that have 200 or more characters: log base 50 of len/200.
The TR special character bonus has been halved (5 to 2.5 time the current points extra, for each char).
The handCombo now uses a power function (x^2), instead of an exponential function (2^x). As a result, the graph doesn't grow as ridiculously large as it did before, though, the limit at 10 combo is still imposed.

* recalculate database promptms (#30)

* Teamvs (#31)

* added teamvs mode

* team typing test file

* Update pull.sh

* Changelog (#32)

Added a changelog, to more easily keep track of versions, and changed the version number in pom.xml.

Co-authored-by: Ashwin Rajesh <ashwinr2k2@gmail.com>
Co-authored-by: Ashwin Rajesh <46510831+VanillaViking@users.noreply.github.com>